mejorando PK y FK de tablas

+++ database/migrations/2021_11_29_160740_create_tbl_provincias_table.php
@@ -16,7 +16,7 @@ class CreateTblProvinciasTable extends Migration
         Schema::create('tbl_provincias', function (Blueprint $table) {
             $table->id('ID_PROVINCIA');
             $table->string('NO_PROVINCIA', 100);
-            $table->integer('FK_DEPARTAMENTO');
+            $table->unsignedBigInteger('FK_DEPARTAMENTO');
             
             $table->foreign('FK_DEPARTAMENTO', 'tbl_departamento_tbl_provincia_fk')->references('ID_DEPARTAMENTO')->on('tbl_departamentos');
         });
