+++ database/migrations/2021_11_29_160744_create_mae_socios_table.php
@@ -16,7 +16,7 @@ class CreateMaeSociosTable extends Migration
         Schema::create('mae_socios', function (Blueprint $table) {
             $table->id('PK_ID_SOCIO');
             $table->char('ES_SOCIO', 1);
-            $table->integer('FK_DIRECCION');
+            $table->unsignedBigInteger('FK_DIRECCION');
             
             $table->foreign('PK_ID_SOCIO', 'mae_persona_tbl_socio_fk')->references('PK_ID_SOCIO')->on('mae_personas');
             $table->foreign('FK_DIRECCION', 'tbl_direccion_tbl_socio_fk')->references('ID_DIRECCION')->on('tbl_direcciones');
