+++ database/migrations/2021_11_29_160751_create_tbl_asistencias_table.php
@@ -17,8 +17,8 @@ class CreateTblAsistenciasTable extends Migration
             $table->id('ID_ASISTENCIA');
             $table->char('ES_SOCIO', 1);
             $table->date('FE_LLEGADA_SOCIO');
-            $table->integer('FK_EVENTO');
-            $table->integer('FK_SOCIO');
+            $table->unsignedBigInteger('FK_EVENTO');
+            $table->unsignedBigInteger('FK_SOCIO');
             
             $table->foreign('FK_EVENTO', 'tbl_evento_tbl_asistencia_fk')->references('ID_EVENTO')->on('tbl_eventos');
             $table->foreign('FK_SOCIO', 'tbl_socio_tbl_asistencia_fk')->references('PK_ID_SOCIO')->on('mae_socios');
