+++ database/migrations/2021_11_29_160752_create_tbl_egresos_table.php
@@ -17,7 +17,7 @@ class CreateTblEgresosTable extends Migration
             $table->id('ID_EGRESO');
             $table->string('NO_EGRESO', 150);
             $table->float('MO_DINERO');
-            $table->integer('FK_SOCIO');
+            $table->unsignedBigInteger('FK_SOCIO');
             
             $table->foreign('FK_SOCIO', 'tbl_socio_tbl_egresos_fk')->references('PK_ID_SOCIO')->on('mae_socios');
         });
